pyproject.toml
[tool.poetry]
name = "hw13"
version = "0.1.1"
description = "GoIT Homework 13, FastAPI email and security"
authors = ["piavik <piavik@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
fastapi-mail = "^1.4.1"
SQLAlchemy = "^2.0.28"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
passlib = { extras = ['bcrypt'], version = "^1.7.4" }
python-multipart = "^0.0.9"
redis = "^5.0.2"
fastapi-limiter = "^0.1.6"
libgravatar = "^1.0.4"
cloudinary = "^1.39.0"

[tool.poetry.group.dev.dependencies]
Faker = "^24.0.0"
requests = "^2.31.0"
sphinx = "^7.2.6"



[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

